snapshot --as-root
Added a flag --as-root to the snapshot command to specify a single directory
path to be the root of the snapshot. This root node is not used for the
snapshot as it is the root, but its contents are explored. This could be useful
to make a snapshot of the contents of a single path, especially if the root
is a device.

The root must always be a single, existing directory node.